qemu's audio subdirectory contains a copy of BSD's sys-queue.h, which
defines a bunch of LIST_ macros.  This makes it difficult to build a
program made partly out of qemu and partly out of the Linux kernel,
since Linux has a different set of LIST_ macros.  It might also cause
trouble when mixing with BSD-derived code.

Under the circumstances it's probably best to rename the versions in
qemu.  The attached patch does this.

committer: Ian Jackson <Ian.Jackson@eu.citrix.com>
git-commit-id: bf4b8549b70e1ed721b372771eaae04e7a5a0f8a