Refactor and fix do_sendkey (Jan Kiszka).
Looking at the sendkey implementation, planning to enhance it with a
hold time argument, I found some potential out-of-bound access and not
very readable code. Here is a fix for the former and a (subjective)
improvement of the latter.

Signed-off-by: Jan Kiszka <jan.kiszka@web.de>


git-svn-id: svn://svn.savannah.nongnu.org/qemu/trunk@4657 c046a42c-6fe2-441c-8c8c-71466251a162

committer: balrog <balrog@c046a42c-6fe2-441c-8c8c-71466251a162>
git-commit-id: 3401c0d95ffb9a9a57093ee002d24d014ffed4f8