Set VNC password from xenstore.
Upstream doesn't appear to have a way to set the password other than
from the monitor, so we introduce this which is a bit messy.

When upstream's new option/config parsing is merged, we should make a
patch for upstream with a config option to read the password out of a
file or fd, and use that.

committer: Ian Jackson <Ian.Jackson@eu.citrix.com>
git-commit-id: 5570f85e1e3b616bae4d5aa6fa727b3a5c3e429a