Revert "ACPI S3 state wake up - change to write flag behaviour"
This reverts commit 7fc722da73707a0965f8b43a67f2ef5fd9a9c143.

Following discussion on xen-devel and elsewhere, we have concluded
that the real problem here was that the S3 resume flag was in the ROM
memory region rather than the EBDA.  This has been fixed in
xen-unstable 18120.

We can therefore keep the whole of the ROM/BIOS area readonly.

Signed-off-by: Ian Jackson <ian.jackson@eu.citrix.com>

committer: Ian Jackson <Ian.Jackson@eu.citrix.com>
git-commit-id: 2609cb7a138cd64ea46ddcf237b2d3be87fcb097