Add a virtual HCI.
This implements most of the logic of a real HCI (at least the pieces
marked as mandatory).  It doesn't support keys, authentication etc.
It works on top of the LMP layer, which is not fully emulated because
software never has direct access to it.


git-svn-id: svn://svn.savannah.nongnu.org/qemu/trunk@5345 c046a42c-6fe2-441c-8c8c-71466251a162

committer: balrog <balrog@c046a42c-6fe2-441c-8c8c-71466251a162>
git-commit-id: 4e38eb5438f254c7f2d588ee15a8045e734143bd