usb-uchi: post-merge: Revert asynchronous completion changes
Previously, during the merge, I resolved the changeset
  2cfc40201be5562a942615452fc2c4ee45d1165c
  usb-uhci: correctly deal with interrupt in asynchronous completion
which came from xen-unstable
  changeset:   12775:60bbcf799384d779c2a561b9d9ba30f28e31d970
  date:        Thu Dec 07 11:52:26 2006 +0000
  files:       tools/ioemu/hw/usb-hid.c tools/ioemu/hw/usb-uhci.c
  description:
  [HVM] qemu mouse: Adds support for USB mouse/tablet status check and
  ...
  And UHC should leave a TD active when receiving NAK and execute this
  incompleted TD in a subseqent frame. UHC only generates an interrupt
  on complete after the TD with ICO bit is completed.
  ...
  Signed-off-by: Xinmei Huang <xinmei.huang@intel.com>
with the new uhci controller in upstream qemu.

However, in the context of the new uhci controller this is a broken
change; it causes the polling of the usb tablet (for example) to cease
after the first two polls.

So I'm reverting it.

committer: Ian Jackson <Ian.Jackson@eu.citrix.com>
git-commit-id: c0acf4d699b13200fbad8f19b6ca89f23eb46317